Move existence check to URLTarget
