+++ tasks/base_tasks.py
@@ -411,24 +411,29 @@ class RepoFileUncompressTask(Task):
     def version(self):
         return 1
 
-    def requires(self):
-        return RepoFile(resource_id=self.task_id, version=self.version(), url=self.get_url())
-
     def get_url(self):
         '''
         Subclasses must override this.
         '''
         raise NotImplementedError('RepoFileUncompressTask must define get_url()')
 
-    def copy_from_repo(self):
-        copyfile(self.input().path,
+    def _copy_from_repo(self, repo_file):
+        copyfile(repo_file.path,
                  '{output}.{extension}'.format(output=self.output().path,
                                                extension=self.compressed_extension))
 
     def run(self):
+        # This is needed with `yield` because `get_url` implementations
+        # might as well depend on input dependencies, so if RepoFile
+        # dependency was specified in `require`, a circular dependency
+        # would arise
+        repo_file = yield RepoFile(resource_id=self.task_id,
+                                   version=self.version(),
+                                   url=self.get_url())
+
         os.makedirs(self.output().path)
         try:
-            self.copy_from_repo()
+            self._copy_from_repo(repo_file)
             self.uncompress()
         except:
             os.rmdir(self.output().path)
@@ -1670,11 +1675,21 @@ def base_downloader(url, output_path):
     urllib.request.urlretrieve(url, output_path)
 
 
+def cdrc_downloader(url, output_path):
+    if 'CDRC_COOKIE' not in os.environ:
+        raise ValueError('This task requires a CDRC cookie. Put it in the `.env` file\n'
+                         'e.g: CDRC_COOKIE=\'auth_tkt="00000000000000000username!userid_type:unicode"\'')
+    shell('wget --header=\'Cookie: {cookie}\' -O {output} {url}'.format(
+        output=output_path,
+        url=url,
+        cookie=os.environ['CDRC_COOKIE']))
+
+
 class RepoFile(Task):
     resource_id = Parameter()
     url = Parameter()
     version = IntParameter()
-    downloader = Parameter(default=base_downloader, significant=False)
+    downloader = Parameter(default='base_downloader', significant=False)
 
     _repo_dir = 'repository'
     _new_file_name = str(uuid.uuid4())
@@ -1692,7 +1707,12 @@ class RepoFile(Task):
         LOGGER.info('Downloading remote file')
         if os.path.isfile(self.output().path):
             os.remove(self.output().path)
-        self.downloader(self.url, self.output().path)
+
+        # As you can't pass functions through Luigi parameters, this
+        # choose of behaviour is needed. A better design would involve
+        # dependencies.
+        downloader = cdrc_downloader if self.downloader == 'cdrc' else base_downloader
+        downloader(self.url, self.output().path)
 
     def _to_db(self, resource_id, version, checksum, url, path):
         LOGGER.info('Storing entry in repository')
@@ -1943,9 +1963,6 @@ class GoogleStorageTask(Task):
             destination_file_name))
 
         blob.download_to_filename(destination_file_name)
-
-    def complete(self):
-        return os.path.exists(self._destination())
 
     def output(self):
         return URLTarget(self._destination())
